refactor: changing files directories
